services/notifications/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.3.0  # For MySQL database interaction
pytest==8.3.2
pytest-mock==3.10.0
requests==2.31.0
python-dotenv==1.0.1
pika  # For RabbitMQ interaction